Refactor: make event directory in root
