+++ src/main/java/world/trecord/controller/notification/NotificationController.java
@@ -13,10 +13,10 @@ import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
 import world.trecord.config.security.CurrentUser;
 import world.trecord.controller.ApiResponse;
 import world.trecord.domain.notification.enumeration.NotificationType;
+import world.trecord.event.sse.SseEmitterService;
 import world.trecord.service.notification.NotificationService;
 import world.trecord.service.notification.response.CheckNewNotificationResponse;
 import world.trecord.service.notification.response.NotificationListResponse;
-import world.trecord.service.sse.SseEmitterService;
 import world.trecord.service.users.UserContext;
 
 import java.time.Duration;
