+++ src/main/java/world/trecord/event/notification/NotificationEvent.java
@@ -1,7 +1,7 @@
-package world.trecord.service.notification;
+package world.trecord.event.notification;
 
-import world.trecord.domain.notification.enumeration.NotificationType;
 import world.trecord.domain.notification.args.NotificationArgs;
+import world.trecord.domain.notification.enumeration.NotificationType;
 
 public record NotificationEvent(Long userToId, Long userFromId, NotificationType type, NotificationArgs args) {
 }
