+++ src/test/java/world/trecord/service/notification/NotificationServiceTest.java
@@ -19,13 +19,13 @@ import world.trecord.domain.record.RecordRepository;
 import world.trecord.domain.userrecordlike.UserRecordLikeRepository;
 import world.trecord.domain.users.UserEntity;
 import world.trecord.domain.users.UserRepository;
+import world.trecord.event.sse.SseEmitterRepository;
 import world.trecord.infra.AbstractContainerBaseTest;
 import world.trecord.infra.IntegrationTestSupport;
 import world.trecord.service.comment.CommentService;
 import world.trecord.service.notification.response.CheckNewNotificationResponse;
 import world.trecord.service.notification.response.NotificationListResponse;
 import world.trecord.service.record.RecordService;
-import world.trecord.service.sse.SseEmitterRepository;
 import world.trecord.service.userrecordlike.UserRecordLikeService;
 
 import java.time.LocalDateTime;
